Move session.state test to its own file
node-http2.test.js contains a 10k-request stress test that takes ~100s
under debug+ASAN and times out when the file's concurrent describe blocks
run alongside it. That timeout is pre-existing on main and unrelated to
this change; placing the new test in its own file lets it be verified
independently.
