+++ test/js/node/http2/node-http2.test.js
@@ -3186,47 +3186,3 @@ it("http2 allowHTTP1 fallback omits the Connection header on a close-delimited r
     server.close();
   }
 });
-
-it("http2 session.state returns an empty object after destroy()", async () => {
-  // node: "An object describing the current status of this Http2Session. If the Http2Session is no
-  // longer usable, an empty object is returned." Property access on the post-destroy value must
-  // not throw.
-  const server = http2.createServer();
-  let serverSession;
-  server.on("session", session => {
-    serverSession = session;
-    session.on("error", () => {});
-  });
-  server.on("stream", stream => {
-    stream.respond({ ":status": 200 });
-    stream.end("ok");
-  });
-  await new Promise(resolve => server.listen(0, resolve));
-  try {
-    const client = http2.connect(`http://localhost:${server.address().port}`);
-    client.on("error", () => {});
-    const { promise, resolve, reject } = Promise.withResolvers();
-    const req = client.request({ ":path": "/" });
-    req.on("error", reject);
-    req.resume();
-    req.on("close", resolve);
-    await promise;
-
-    expect(typeof client.state.effectiveLocalWindowSize).toBe("number");
-
-    client.destroy();
-    const clientState = client.state;
-    expect(client.destroyed).toBe(true);
-    expect(clientState).toEqual({});
-    expect(() => clientState.effectiveLocalWindowSize).not.toThrow();
-
-    expect(serverSession).toBeDefined();
-    serverSession.destroy();
-    const serverState = serverSession.state;
-    expect(serverSession.destroyed).toBe(true);
-    expect(serverState).toEqual({});
-    expect(() => serverState.effectiveLocalWindowSize).not.toThrow();
-  } finally {
-    server.close();
-  }
-});
